+++ src/PhpSpreadsheet/Worksheet/Worksheet.php
@@ -13,9 +13,9 @@ use PhpOffice\PhpSpreadsheet\Chart\Chart;
 use PhpOffice\PhpSpreadsheet\Collection\Cells;
 use PhpOffice\PhpSpreadsheet\Collection\CellsFactory;
 use PhpOffice\PhpSpreadsheet\Comment;
+use PhpOffice\PhpSpreadsheet\DefinedName;
 use PhpOffice\PhpSpreadsheet\Exception;
 use PhpOffice\PhpSpreadsheet\IComparable;
-use PhpOffice\PhpSpreadsheet\NamedRange;
 use PhpOffice\PhpSpreadsheet\ReferenceHelper;
 use PhpOffice\PhpSpreadsheet\RichText\RichText;
 use PhpOffice\PhpSpreadsheet\Shared;
@@ -797,9 +797,9 @@ class Worksheet implements IComparable
     public function rebindParent(Spreadsheet $parent)
     {
         if ($this->parent !== null) {
-            $namedRanges = $this->parent->getNamedRanges();
-            foreach ($namedRanges as $namedRange) {
-                $parent->addNamedRange($namedRange);
+            $definedNames = $this->parent->getDefinedNames();
+            foreach ($definedNames as $definedName) {
+                $parent->addDefinedName($definedName);
             }
 
             $this->parent->removeSheetByIndex(
@@ -1192,11 +1192,11 @@ class Worksheet implements IComparable
         // Named range?
         if (
             (!preg_match('/^' . Calculation::CALCULATION_REGEXP_CELLREF . '$/i', $pCoordinate, $matches)) &&
-            (preg_match('/^' . Calculation::CALCULATION_REGEXP_NAMEDRANGE . '$/i', $pCoordinate, $matches))
+            (preg_match('/^' . Calculation::CALCULATION_REGEXP_DEFINEDNAME . '$/i', $pCoordinate, $matches))
         ) {
-            $namedRange = NamedRange::resolveRange($pCoordinate, $this);
+            $namedRange = DefinedName::resolveName($pCoordinate, $this);
             if ($namedRange !== null) {
-                $pCoordinate = $namedRange->getRange();
+                $pCoordinate = $namedRange->getValue();
 
                 return $namedRange->getWorksheet()->getCell($pCoordinate, $createIfNotExists);
             }
@@ -1292,11 +1292,11 @@ class Worksheet implements IComparable
         // Named range?
         if (
             (!preg_match('/^' . Calculation::CALCULATION_REGEXP_CELLREF . '$/i', $pCoordinate, $matches)) &&
-            (preg_match('/^' . Calculation::CALCULATION_REGEXP_NAMEDRANGE . '$/i', $pCoordinate, $matches))
+            (preg_match('/^' . Calculation::CALCULATION_REGEXP_DEFINEDNAME . '$/i', $pCoordinate, $matches))
         ) {
-            $namedRange = NamedRange::resolveRange($pCoordinate, $this);
+            $namedRange = DefinedName::resolveName($pCoordinate, $this);
             if ($namedRange !== null) {
-                $pCoordinate = $namedRange->getRange();
+                $pCoordinate = $namedRange->getValue();
                 if ($this->getHashCode() != $namedRange->getWorksheet()->getHashCode()) {
                     if (!$namedRange->getLocalOnly()) {
                         return $namedRange->getWorksheet()->cellExists($pCoordinate);
@@ -2564,10 +2564,10 @@ class Worksheet implements IComparable
      */
     public function namedRangeToArray($pNamedRange, $nullValue = null, $calculateFormulas = true, $formatData = true, $returnCellRef = false)
     {
-        $namedRange = NamedRange::resolveRange($pNamedRange, $this);
+        $namedRange = DefinedName::resolveName($pNamedRange, $this);
         if ($namedRange !== null) {
             $pWorkSheet = $namedRange->getWorksheet();
-            $pCellRange = $namedRange->getRange();
+            $pCellRange = $namedRange->getValue();
 
             return $pWorkSheet->rangeToArray($pCellRange, $nullValue, $calculateFormulas, $formatData, $returnCellRef);
         }