+++ CHANGELOG.md
@@ -11,14 +11,36 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 - Implemented Page Order for Xlsx and Xls Readers, and provided Page Settings (Orientation, Scale, Horizontal/Vertical Centering, Page Order, Margins) support for Ods, Gnumeric and Xls Readers [#1559](https://github.com/PHPOffice/PhpSpreadsheet/pull/1559)
 - Implementation of the Excel `LOGNORM.DIST()`, `NORM.S.DIST()`, `GAMMA()` and `GAUSS()` functions. [#1588](https://github.com/PHPOffice/PhpSpreadsheet/pull/1588)
+- Named formula implementation, and improved handling of Defined Names generally [#1535](https://github.com/PHPOffice/PhpSpreadsheet/pull/1535)
+  - Defined Names are now case-insensitive
+  - Distinction between named ranges and named formulae
+  - Correct handling of union and intersection operators in named ranges
+  - Correct evaluation of named range operators in calculations
+  - fix resolution of relative named range values in the calculation engine; previously all named range values had been treated as absolute.
+  - Calculation support for named formulae
+  - Support for nested ranges and formulae (named ranges and formulae that reference other named ranges/formulae) in calculations
+  - Introduction of a helper to convert address formats between R1C1 and A1 (and the reverse)
+  - Proper support for both named ranges and named formulae in all appropriate Readers
+    - **Xlsx** (Previously only simple named ranges were supported)
+    - **Xls** (Previously only simple named ranges were supported)
+    - **Gnumeric** (Previously neither named ranges nor formulae were supported)
+    - **Ods** (Previously neither named ranges nor formulae were supported)
+    - **Xml** (Previously neither named ranges nor formulae were supported)
+  - Proper support for named ranges and named formulae in all appropriate Writers
+    - **Xlsx** (Previously only simple named ranges were supported)
+    - **Xls** (Previously neither named ranges nor formulae were supported) - Still not supported, but some parser issues resolved that previously failed to differentiate between a defined name and a function name
+    - **Ods** (Previously neither named ranges nor formulae were supported)
 
 ### Changed
 
 - Improve Coverage for ODS Reader [#1545](https://github.com/phpoffice/phpspreadsheet/pull/1544)
+- Named formula implementation, and improved handling of Defined Names generally [#1535](https://github.com/PHPOffice/PhpSpreadsheet/pull/1535)
+  - fix resolution of relative named range values in the calculation engine; previously all named range values had been treated as absolute.
 
 ### Deprecated
 
-- Nothing.
+-   **IMPORTANT NOTE:** This Introduces a **BC break** in the handling of named ranges. Previously, a named range cell reference of `B2` would be treated identically to a named range cell reference of `$B2` or `B$2` or `$B$2` because the calculation engine treated then all as absolute references. These changes "fix" that, so the calculation engine now handles relative references in named ranges correctly.
+  This change that resolves previously incorrect behaviour in the calculation may affect users who have dynamically defined named ranges using relative references when they should have used absolute references.
 
 ### Removed
 
@@ -40,8 +62,6 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 - Borders were not complete on rowspanned columns using HTML reader [#1473](https://github.com/PHPOffice/PhpSpreadsheet/pull/1473)
 
 ### Changed
-
-- nothing
 
 ## 1.14.0 - 2020-06-29